+++ tests/Identity/Model/AccountTest.php
@@ -126,6 +126,40 @@ class AccountTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @test
+     * @depends it_should_create_new_account
+     *
+     * @param Account $account
+     */
+    public function it_should_change_password($account)
+    {
+        $updated_time = new DateTime('tomorrow');
+        $newPassword = new HashedPassword('new_valid_hashed_password');
+
+        DateTime::setTestDateTime($updated_time);
+        $account->changePassword($newPassword);
+        DateTime::clearTestDateTime();
+
+        $this->assertEquals(1, count($account->releaseEvents()));
+        $this->assertEquals($newPassword, $account->getPassword());
+        $this->assertEquals($updated_time, $account->getUpdatedAt());
+    }
+
+    /**
+     * @test
+     */
+    public function it_should_not_change_password_when_locked()
+    {
+        $this->setExpectedException(AccountIsLocked::class);
+
+        $account = Account::create($this->accountId, $this->email, $this->password);
+        $account->lockHard();
+        $newPassword = new HashedPassword('new_valid_hashed_password');
+
+        $account->changePassword($newPassword);
+    }
+
+    /**
+     * @test
      *
      * @return Account
      */